chore: separate serstate out
