+++ vm/rust/src/juno_state_reader.rs
@@ -1,5 +1,4 @@
 use blockifier::execution::contract_class::{ContractClass, NativeContractClassV1};
-use blockifier::state::cached_state::StorageEntry;
 use blockifier::state::errors::StateError;
 use blockifier::{
     execution::contract_class::{
@@ -15,12 +14,11 @@ use cairo_native::{
 };
 use libloading::Library;
 use once_cell::sync::Lazy;
-use serde::{Deserialize, Serialize};
+use serde::Deserialize;
 use starknet_api::core::{ClassHash, CompiledClassHash, ContractAddress, Nonce};
 use starknet_api::state::StorageKey;
 use starknet_types_core::felt::Felt;
 use std::cell::RefCell;
-use std::collections::HashMap;
 use std::{
     ffi::{c_char, c_uchar, c_void, CStr},
     fs,
@@ -28,6 +26,8 @@ use std::{
     slice,
     sync::Mutex,
 };
+
+use crate::serstate::SerState;
 
 extern "C" {
     fn JunoFree(ptr: *const c_void);
@@ -56,59 +56,6 @@ struct CachedContractClass {
 
 static CLASS_CACHE: Lazy<Mutex<SizedCache<ClassHash, CachedContractClass>>> =
     Lazy::new(|| Mutex::new(SizedCache::with_size(128)));
-
-#[derive(Default, Serialize, Deserialize)]
-pub struct SerState {
-    // Could have been wrapped around with felt
-    storage: HashMap<StorageEntry, [u8; 32]>,
-    nonce: HashMap<ContractAddress, Nonce>,
-    class_hash: HashMap<ContractAddress, ClassHash>,
-    contract_class: HashMap<ClassHash, String>,
-}
-
-impl StateReader for SerState {
-    fn get_storage_at(
-        &self,
-        contract_address: ContractAddress,
-        key: StorageKey,
-    ) -> StateResult<Felt> {
-        // might have to deal with
-        let bytes = self
-            .storage
-            .get(&(contract_address, key))
-            .expect("no storage");
-        Ok(Felt::from_bytes_be(bytes))
-    }
-
-    fn get_nonce_at(&self, contract_address: ContractAddress) -> StateResult<Nonce> {
-        Ok(*self
-            .nonce
-            .get(&contract_address)
-            .expect("no nonce for this"))
-    }
-
-    fn get_class_hash_at(&self, contract_address: ContractAddress) -> StateResult<ClassHash> {
-        Ok(*self
-            .class_hash
-            .get(&contract_address)
-            .expect("no nonce for this"))
-    }
-
-    fn get_compiled_contract_class(&self, class_hash: ClassHash) -> StateResult<ContractClass> {
-        // Passing along version information could make this a V1 and V1Native test
-        let json_str = self
-            .contract_class
-            .get(&class_hash)
-            .expect("request non existed class");
-        Ok(class_info_from_json_str(json_str, class_hash)
-            .expect("decoding class went wrong")
-            .contract_class())
-    }
-
-    fn get_compiled_class_hash(&self, _class_hash: ClassHash) -> StateResult<CompiledClassHash> {
-        unimplemented!()
-    }
-}
 
 pub struct JunoStateReader {
     pub handle: usize, // uintptr_t equivalent
